commands: move status code to status.rs
